Created an api route for updating listings.

+++ Api/controller/listing.controller.js
@@ -1,28 +1,57 @@
+import mongoose from 'mongoose';
 import Listing from "../models/listing.model.js";
-
-export const create= async(req,res,next)=>{
-    try {
-        const listing=await Listing.create(req.body);
-        res.status(201).json(listing);
-    } catch (error) {
-        next(error);
-    }
-}
-export const deleteList=async(req,res,next)=>{
-    const listing = await Listing.findById(req.params.id);
+import  {errorHandler}  from '../utils/error.js';
+export const create = async (req, res, next) => {
+  try {
+    const listing = await Listing.create(req.body);
+    res.status(201).json(listing);
+  } catch (error) {
+    next(error);
+  }
+};
+export const deleteList = async (req, res, next) => {
+  // Check if req.params.id is a valid ObjectId
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).send('Invalid ObjectId');
+  }
+  const listing = await Listing.findById(req.params.id);
 
   if (!listing) {
-    return next(errorHandler(404, 'Listing not found!'));
+    return next(errorHandler(404, "Listing not found!"));
   }
 
   if (req.user.id !== listing.userRef) {
-    return next(errorHandler(401, 'You can only delete your own listings!'));
+    return next(errorHandler(401, "You can only delete your own listings!"));
   }
 
   try {
     await Listing.findByIdAndDelete(req.params.id);
-    res.status(200).json('Listing has been deleted!');
+    res.status(200).json("Listing has been deleted!");
   } catch (error) {
     next(error);
   }
-}
+};
+export const updateList = async (req, res, next) => {
+   // Check if req.params.id is a valid ObjectId
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).send('Invalid ObjectId');
+    }
+    const listing = await Listing.findById(req.params.id);
+    if (!listing) {
+      return next(errorHandler(404, 'Listing not found!'));
+    }
+    if (req.user.id !== listing.userRef) {
+      return next(errorHandler(401, 'You can only update your own listings!'));
+    }
+  
+    try {
+      const updatedListing = await Listing.findByIdAndUpdate(
+        req.params.id,
+        req.body,
+        { new: true }
+      );
+      res.status(200).json(updatedListing);
+    } catch (error) {
+      next(error);
+    }
+  };
